refactor: follow project structure conventions
