+++ .env.example
@@ -2,4 +2,3 @@ DB_USER=DATABASE_USER
 DB_PASSWORD=DATABASE_PASSWORD
 DB_NAMESPACE=lex
 DB_DATABASE=production
-DB_CONNECTION_URL=http://localhost:8000
