+++ main.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/crux-bphc/lex/internal"
+	"github.com/crux-bphc/lex/internal/routes"
 	"github.com/gin-contrib/cors"
 	"github.com/gin-contrib/location"
 	"github.com/gin-gonic/gin"
@@ -12,7 +12,9 @@ func main() {
 	router.Use(cors.Default())
 	router.Use(location.Default())
 
-	internal.RegisterImpartusRoutes(router)
+	router.SetTrustedProxies(nil)
+
+	routes.RegisterImpartusRoutes(router)
 
 	router.Run(":3000")
 }
